moderator roll added

+++ app/controllers/posts_controller.rb
@@ -1,7 +1,7 @@
 class PostsController < ApplicationController
 
 before_action :require_sign_in, except: :show
-before_action :authorize_user, except: [:show, :new, :create]
+# before_action :authorize_user, except: [:show, :new, :create]
 
   def show
     @post = Post.find(params[:id])
@@ -26,33 +26,51 @@ before_action :authorize_user, except: [:show, :new, :create]
    end
 
   def edit
-    @post = Post.find(params[:id])
+    post = Post.find(params[:id])
+    if current_user == post.user || current_user.admin? || current_user.moderator?
+      @post = Post.find(params[:id])
+    else
+      flash[:alert] = "You must be an admin to do that."
+      redirect_to [post.topic, post]
+    end
   end
 
 
   def update
-       @post = Post.find(params[:id])
-       @post.assign_attributes(post_params)
+        post = Post.find(params[:id])
+       if current_user == post.user || current_user.admin? || current_user.moderator?
+         @post = Post.find(params[:id])
+         @post.assign_attributes(post_params)
 
-       if @post.save
-         flash[:notice] = "Post was updated."
-         redirect_to [@post.topic, @post]
-       else
-         flash.now[:alert] = "There was an error saving the post. Please try again."
-         render :edit
-       end
+         if @post.save
+           flash[:notice] = "Post was updated."
+           redirect_to [@post.topic, @post]
+         else
+           flash.now[:alert] = "There was an error saving the post. Please try again."
+           render :edit
+         end
+      else
+        flash[:alert] = "You must be an admin to do that."
+        redirect_to [post.topic, post]
+      end
      end
 
      def destroy
-     @post = Post.find(params[:id])
+       post = Post.find(params[:id])
+       if current_user == post.user || current_user.admin?
+         @post = Post.find(params[:id])
 
-     if @post.destroy
-       flash[:notice] = "\"#{@post.title}\" was deleted successfully."
-       redirect_to @post.topic
-     else
-       flash.now[:alert] = "There was an error deleting the post."
-       render :show
-     end
+         if @post.destroy
+           flash[:notice] = "\"#{@post.title}\" was deleted successfully."
+           redirect_to @post.topic
+         else
+           flash.now[:alert] = "There was an error deleting the post."
+           render :show
+         end
+       else
+         flash[:alert] = "You must be an admin to do that."
+         redirect_to [post.topic, post]
+       end
    end
 
    private
@@ -60,12 +78,12 @@ before_action :authorize_user, except: [:show, :new, :create]
      params.require(:post).permit(:title, :body)
    end
 
-   def authorize_user
-     post = Post.find(params[:id])
-     unless current_user == post.user || current_user.admin?
-       flash[:alert] = "You must be an admin to do that."
-       redirect_to [post.topic, post]
-     end
-   end
+   # def authorize_user
+   #   post = Post.find(params[:id])
+   #   unless current_user == post.user || current_user.admin?
+   #     flash[:alert] = "You must be an admin to do that."
+   #     redirect_to [post.topic, post]
+   #   end
+   # end
 
 end
